fea: Se continua con migración a microstream iniciando cambio de EJB y Queries.

+++ sms-builder-web/src/main/java/co/edu/utp/gia/sms/dtos/ReferenciaDTO.java
@@ -71,7 +71,7 @@ public class ReferenciaDTO implements Serializable {
     }
 
     private void evaluarSeleccion() {
-        seleccionada = (referencia.getFiltro().intValue() | etapa) > etapa;
+        seleccionada = (referencia.getFiltro() | etapa) > etapa;
     }
 
     /**
